move map between head and items list
